feature(blackjack): Bank 내부의 베팅 금액을 Money 라는 value class 로 래핑

+++ src/main/kotlin/blackjack/ui/ConsoleOutput.kt
@@ -62,7 +62,7 @@ class ConsoleOutput {
             println("## 최종 수익")
             profits.forEach {
                 print("${it.participant.name}: ")
-                println(it.profit)
+                println(it.profit.toInt())
             }
         }
     }
